requirements.txt
annotated-types==0.7.0
anthropic==0.72.0
anyio==4.11.0
asgiref==3.8.1
certifi==2025.10.5
charset-normalizer==3.4.4
distro==1.9.0
Django==5.0.6
docstring_parser==0.17.0
et_xmlfile==2.0.0
gunicorn==22.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.11.1
numpy==2.4.0
openpyxl==3.1.5
packaging==24.1
pandas==2.3.3
pillow==12.0.0
psycopg2-binary==2.9.11
python-dotenv==1.0.1
pydantic==2.12.3
pydantic_core==2.41.4
python-dateutil==2.9.0.post0
pytz==2025.2
reportlab==4.4.7
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
